replaceRevive: Copy remotedev-serialize's tests, commented out.
We'll soon be inlining remotedev-serialize, and we'll want to at
least maintain the same test coverage the code had when it lived
there.

At first, the tests will just import remotedev-serialize's code and
run that; then, we'll switch it over to testing that code as it
appears in our project.

Start with the tests and snapshots commented out; then, once we've
made a few changes in the next few commits, they'll be able to run.