replaceRevive tests [nfc]: Update snapshots.
The new code is obtained by uncommenting the tests in
replaceRevive-test.js and running them (then re-commenting
everything for this commit; we'll commit the uncommenting soon).

I'm not sure why we suddenly have to use `\\` instead of `\` to
escape quotes in the JSON. Other than that, I don't see any
discrepancies.