narrow: Check IDs as well as emails in isNarrowValid.
We use this function to avoid trying to load UI for a narrow where
the UI is doomed to crash because we're lacking basic data about it:
in particular, a PM conversation where one of the putative
participants is a user we have no data for.  One way that can happen
is if a user has changed their email address, and either we learn
about the narrow by their old email address, or we still have them
by their old email address and learn about the narrow by their new one.

In the course of migrating code from using the emails in a
PM-conversation narrow to using the user IDs, we're about to start
doing so with code that will look the users up in our data
structures to get further information like full name or avatar.
So, in the spirit of "crunchy shell", we should apply the same checks
here to the user IDs.

Then at the end of this migration series, when we've converted
everything else that consumes a PM narrow so that it uses IDs
instead of emails, we can drop the email part here because it'll no
longer be possible for it to matter.