sharing: Identify PM recipients by ID alone.
Another step in the conversion to the UserItemById component.

This one is particularly nice because it saves us from keeping
around a list of `User` objects as local state.  That means we
automatically fix a whole class of glitch-level bugs that this
code must have had, where when some details about a user changed
after you'd already chosen to send your sharing message to them,
the information we showed about them wouldn't update as normal.