webview: Fix race when setting scroll position
Setting innerHTML triggers an asynchronous render of the DOM. On iOS
this reliably happens _after_ the code that comes immediately after the
innerHTML set. This means that if we calculate a scroll position right
after a content update, we can get the wrong scrollHeight and present
the wrong position.

A nested requestAnimationFrame callback reliably runs code after the
scrollHeight is updated, so use that to get the accurate scroll target.

References:
https://twitter.com/rauschma/status/1288868746682081285
https://macarthur.me/posts/when-dom-updates-appear-to-be-asynchronous

Test Plan: Manually loaded streams on iPhone 12 (iOS 14.3)

Fixes: #4357