webview [nfc]: Make `platformOS` a global again.
This will let us conveniently refer to it from other parts of the
in-webview JS code.

This effectively reverts much of 1d9c9e1f6.  As an alternate solution
to the problem that solved, we use a Flow `declare` statement to give
the type-checker the same information we were giving by making this
value a parameter.