narrow [nfc]: Use user IDs instead of emails in apiNarrowOfNarrow.
Well, use one to get the other: we still need emails to send to the
server, until we drop support for servers older than 2.1.  Also add
a TODO(server-2.1) comment to help us find that when we do.

But with this change, we don't use the emails *from the Narrow
value* -- which means we're now all clear to remove the emails
from Narrow values entirely.