misc [nfc]: Remove `groupItemsById`.
We just removed the last use of this in the previous commit.

It's one of our several helper functions that hasn't yet grown a
jsdoc. This function is quite short and simple, though, so I don't
think it needs to have its own definition, with or without a jsdoc.
Inlining the logic means one less thing to click through to find out
what a piece of code is doing.