drafts: Fix broken migration of narrow keys.
As the comment in migration 19 said, we were doing something a bit
risky there in order to avoid duplicating the then-current version
of `keyFromNarrow`.  We simply invoked `keyFromNarrow` directly --
which called for being careful about migrations for any future change
we made to `keyFromNarrow`.

Then a bit later we changed `keyFromNarrow` quite dramatically, and
didn't pay attention to migrations at all.  Oops.

Specifically, in cf4207f47 we changed the `Narrow` type to have quite
a different structure.  That's the type that `keyFromNarrow` accepts
-- so the result is that `keyFromNarrow` stopped working on the old
values.  Which are exactly what `JSON.parse` is going to supply us
when this migration is run.  The result is probably that if the user
has any drafts when they take an upgrade that spans migration 19
and commit cf4207f47, the app will crash on startup.

We'd probably have found this bug at the alpha stage when next making
a release, so it wouldn't have made it to users other than ourselves.
Better to fix ASAP, though.

The plan for migration 19, as described here:
  https://github.com/zulip/zulip-mobile/pull/4339#discussion_r543603022
knowingly involved breaking the basic rule of migrations: don't invoke
any of the app's model code, because it will change.  I figured it was
OK because there were changes coming up soon which I'd planned out how
I'd add migrations for.  Then later the same week I sent #4346, which
followed that plan... and I hadn't thought through this particular
interaction, and it introduced this bug.  So this is a fresh lesson in
the basic rule of migrations: don't invoke any of the app's model code,
because it will change.

See discussion where we found this bug:
  https://github.com/zulip/zulip-mobile/pull/4382#discussion_r552283858