PresenceStatusIndicator [nfc]: Add `MaybeOpaqueBackgroundWrapper`.
This commit and the next few aim to make it very easy to see what's
being controlled by `useOpaqueBackground`. By the end of this small
series, the answer will be "`useOpaqueBackground` controls whether a
single set of styles is used, or not used, in a single `View`" --
which is simpler than in the current code, which repeats a somewhat
more complicated condition a few times.

Start by consolidating some of the logic for conditionally using the
opaque background on the status indicator, so that we don't have to
repeat as much code in each of the `status` cases.