unread: Add new getUnreadStreams selector.
This provides a data structure that's just like the one we're going
to switch to maintaining as the state internally.  We'll migrate
consumers to use this one, and then we can change out the state and
reducer underneath it.

As we migrate consumers to this selector, it may initially cause a
certain performance regression: when the "unread" data structure gets
updated, we'll promptly be making this (reformatted) copy of it, as
the selectors that use it go to look for the new data.  But that extra
work will then get eliminated when we switch to this being the data
structure that we maintain in the first place.