unread tests: Delete a spurious test.
Getting the same message ID twice just shouldn't happen -- it'd be
a break in the contract of the server API.  If some server bug does
ever cause it to happen, and we don't notice and just include the
additional message ID here, nothing particularly bad will happen;
at worst a glitch where the unread counts are too high.

Meanwhile, maintaining this data structure is a hot path, so it
won't make sense for us to defensively check for such a server bug.
Take out the test that checks for it.