react-navigation libdef: Fix `getRootState`.
It seems that `PossiblyStaleNavigtionState` [1] is the Flow libdef's
version of what appears in the TypeScript as
`PartialState<NavigationState>` [2].

In contrast with the Flow libdef, the TypeScript has `getRootState`
returning the non-"partial", non-"possibly stale" thing, i.e.,
`NavigationState` [3]. That suggests we should consider making the
Flow libdef say the same thing, which also has the name
`NavigationState`.

React Navigation does have a doc on "Partial state objects" [4].
Based on that, Greg points out [5] that a "stale" nav state is
accepted as input to various React Navigation options, but a
non-stale state is always the output -- suggesting that that the
return value of something like `getRootState` should surely not be
stale.

Some logging confirms that there aren't prominent cases where
`stale` is `true` or missing. (Any such cases would suggest that we
should keep the `PossiblyStaleNavigationState` return type.)

See discussion [6].

[1] https://github.com/flow-typed/flow-typed/blob/master/definitions/npm/%40react-navigation/native_v5.x.x/flow_v0.104.x-/native_v5.x.x.js#L526
[2] https://github.com/react-navigation/react-navigation/blob/main/packages/routers/src/types.tsx#L58
[3] https://github.com/react-navigation/react-navigation/blob/main/packages/core/src/types.tsx#L471
[4] https://reactnavigation.org/docs/navigation-state/#partial-state-objects
[5] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/React.20Navigation.20v5.3A.20.60PossiblyStaleNavigationState.60/near/1097141
[6] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/React.20Navigation.20v5.3A.20.60PossiblyStaleNavigationState.60/near/1096924