prettier: Ignore flow-typed/@react-navigation.
As Greg points out [1], it's best to smoothen the path to making a
PR to the FlowTyped repo with the changes we're about to make to
this libdef. That means not converting it to our own style, which
means telling Prettier and ESLint to ignore these files.

ESLint already ignores them because of the `**/flow-typed/**`
line in `.eslintignore`.

[1] https://github.com/zulip/zulip-mobile/pull/4393#discussion_r556853806