user types: Use UserId for PM recipients.
This PmRecipientUser type is one we reuse in some data structures
we construct ourselves, in particular outbox messages.  So to
tighten its type, we need to propagate the fact that these values
are user IDs through the various code that handles them on their
way there, in particular the Narrow type and its constructors.