deps: Point to PR commit for @react-native-community/cameraroll.
As Greg suggests [1], point to the commit at the tip of my PR
react-native-cameraroll/react-native-cameraroll#264.

Once a version of @react-native-community/cameraroll with official
RN v0.63 support (including as declared in `peerDependencies`) is
released to NPM, we should go back to specifying a version range,
making sure that that range doesn't include versions before that
update.

We expect this will happen relatively soon; the project does seem to
be actively merging PRs, and I recently left a ping on the issue.
But we'd like to start using RN v0.63 in our project ASAP, so we use
this strategy in the meantime.

Sometimes packages have a crucial build/packaging step that gets
done as part of releasing a version to NPM. Such a step would likely
make it impossible to point to a commit on GitHub and have the
package still work. But I've verified that it does indeed work when
we point to the commit on GitHub, so there's no problem there.

[1] https://github.com/zulip/zulip-mobile/pull/4420#issuecomment-763973831