UserItem [nfc]: Drop a spurious "is email empty" condition.
This is pretty dubious logic that, to the extent it's possible for
it to have any effect at all, makes this component's interface more
complicated than it needs to be.

Happily, this condition should never be able to trigger:

 * Most callers of `UserItemRaw` get here through `UserItem`,
   which looks up a real user's data.  Real users don't have the
   empty string (let alone some other falsy value, which wouldn't
   be well-typed) as their "email".

 * The one other caller is `PeopleAutocomplete`.  That one passes
   a number of real users, and possibly a couple of pseudo-users
   generated by `getAutocompleteSuggestion`.  Those two pseudo-users
   have non-empty "emails" too: they both read "(Notify everyone)".

So nothing could have been relying on this quirk, and we can just
cut it out.