ZulipStatusBar [nfc]: Un-nest `StatusBar` from `View`.
React Native's `StatusBar` component is unusual in that its `render`
method always returns null [1]. The component isn't meant to
participate in the spatial layout of the UI elements. Instead, the
API works more like this: Mount a `StatusBar` component in order to
tell something new to the native status bar API; e.g., what colors
and animation you want to use.

This commit is therefore NFC: the `StatusBar` still gets mounted in
all cases where it was getting mounted before, and nothing changes
in the spatial layout of the UI elements.

The `<View style={style} />` is doing an important job: it covers
the "safe area" at the top of the screen (which includes the status
bar), so that our content doesn't accidentally overlap with it. But
`ZulipStatusBar` already has a very different job, which is to
configure the status bar in kind of subtle ways, and depending on
the platform. I'd like to pull the safe-area handling out of
`ZulipStatusBar` to give it one less thing to worry about; that will
happen in the rest of this series of commits.

[1] https://github.com/facebook/react-native/blob/v0.63.4/Libraries/Components/StatusBar/StatusBar.js#L497