store tests: Fix "all keys are in config" test; drop an export.
This test never actually quite checked the thing it said it did:
if you read the code, it was saying the converse, that every key
that *was* listed in the store config was indeed the name of some
reducer.  Then also checking that the lengths matched... but
nothing ensured that the keys listed in the store config are
distinct, so if we were to accidentally have a duplicate then we
could also have one missing, and this test wouldn't notice.

Fix the test, and while we're at it also get the list of keys
straight from the behavior of the overall reducer -- from the state
it returns on initialization -- rather than from the internals of
its implementation.  This lets us stop exporting that bit of its
internals, too.