msglist tests [nfc]: Rename test file to follow our convention.
All our other unit tests live in a file whose name ends in "-test.js",
typically "foo-test.js" where the code being tested is "foo.js".

One reason for this is that it's common for editors and IDEs to
identify a file in the UI primarily by its basename, i.e. the part of
its path after the last slash.  So when the test has the same basename
as the code it's testing, like "rewriteHtml.js" and "rewriteHtml.js",
they tend to collide with each other in the UI.

But more simply: when 92 out of 93 examples follow the same
convention uniformly, having a 93rd that's different is an
unhelpful complication.  There's nothing gained by this exception,
so make it like the rest.