logging: Set `undefined` for server version tags if server version absent.
Instead of omitting them.

We're about to start keeping these tags on the global scope, which
means we have to listen to and propagate all changes in the server
version, including changes where the server version disappears. It's
been reported [1] that passing `undefined` as a tag's value in
`setTags` will clear the tag from the current scope, and I've
verified this experimentally.

[1] https://github.com/getsentry/sentry-javascript/pull/3108#issue-534072956