podfile: Follow RN update to make Podfile not depend on path to RN.
Done to follow the template-app change in
facebook/react-native@cd13c99d0 [1], on the path to the RN v0.64
upgrade.

Greg points out [2] that a later commit,
facebook/react-native@e599d6c5d, suggests that the
`config["reactNativePath"]` form in the earlier commit is actually
probably broken. So, we squash that in.

That later commit isn't very well explained, but it looks like their
claim for the change being needed starts by mentioning
@react-native-community/cli@e949e234b, which was released in v4.10.1
of @react-native-community/cli. `yarn why` tells me that we're on
v4.14.0 of that library, so it seems fine and good to take this.

[1] It actually looks like we maybe could/should have taken this
    change when we did the RN v0.63 upgrade; it was backported to
    v0.63 in facebook/react-native@696fb5588. Looks like we didn't;
    oops.

[2] https://github.com/zulip/zulip-mobile/pull/4887#discussion_r682295252