sharing: Fix bug when app is launched from history.
This was required because of the following bug:
1. share from any app to zulip.
2. cancel the share.
3. return back to zulip by selecting it from minimized apps.
4. now go back to the previous app used to share to zulip.
5. share something to zulip again.
6. the sharing screen will not be visible.

Intended behavior: Sharing screen should be visible in step 6

This bug was possibly introduced in:

70e9ad9d47ae2e192558cb4caa9ae8781251b926

The condition checking `FLAG_ACTIVITY_LAUNCHED_FROM_HISTORY` in the code
that commit introduces, checks MainActivity's intent (intent the app was
originally started with). Instead it should check the intent handleSend
was called with.

This caused the intents coming via `onNewIntent` to be ignored.