types: Make object types with indexer properties exact, in custom libdefs.
Specifically, the libdefs in flow-typed but not in flow-typed/npm.

Done with the same find-and-replace strategy as in a recent commit.

Here, there were substantially more *explicitly* inexact object
types with indexer properties. It doesn't make a lot of sense why
one would want an inexact object with an indexer property, whether
implicitly or explicitly inexact. So, make those exact too.