webview: Don't show actionsheet for PM header long-press
Previously, long-pressing a PM header in the "All Messages" (or similar)
view would pop up a actionsheet, which only showed the name of the PM
and a "Cancel" button. This was pretty silly, and makes the ActionSheet
code more complicated than it needs to be, so we'll just pop up a toast
showing the name of the PM instead. This is useful in the case of a PM
where the title is long enough that it's truncated in the UI.

The code for the toast text is taken from `getActionSheetTitle` in
src/message/messageActionSheet.js - it's duplicated in this commit for
type safety reasons, but the duplication will be removed in the next
commit.