ActionSheets: Remove message arg from constructHeaderActionButtons
This changes the constructHeaderActionButtons function to take a stream
and topic, instead of a message, since that is conceptually closer to
what it's doing.

I haven't routed this through to showHeaderActionButton yet, since all
of the button callbacks currently still require a Message, but that's
the goal in the future.