accountsReducer tests: Remove a now-redundant test.
It looks like this test was specifically aimed at the behavior of
taking partial information in the "initial account" and filling it
out on login success. But that specific behavior doesn't exist
anymore, because we've recently stopped storing partial login
information in Redux.

In fact, the coverage this test provides is subsumed by the "on
login, if account does exist, ..." test below. So, remove this test
as, in effect, a duplicate.