AccountItem: Tweak the layout of "Signed out" text.
The previous position of this text looked just a little odd: to my
eye, it didn't play nicely with the vertical centering of the
trash-can icon. Now, it joins the existing left-aligned labels as
the last child of the `styles.details` View.

I'm not sure I've arrived at the best layout choice, but at least
there can be no question about vertical centering. It also makes the
code slightly easier to reason about, since it mostly fits into an
existing pattern: it's just like the existing left-aligned labels,
except it's italic instead of bold (which I like).