events [nfc]: Inline eventsToActions.
This makes manifest a bug!  We've been getting the Redux state at the
start of this loop, and using it to process all the events... which
means that for later events in the batch, it may be out of date.

As mentioned in a previous commit of this series where we eliminated
dispatchOrBatch, it's rare while the app is running and online to have
a batch of more than one event anyway.  So the impact of this bug has
probably been small.  But a larger batch may be realistic if for
example the device was off the network and then came back on.

Anyway, we'll fix that bug next.