lightbox: Fix sharing from lightbox screen for Android.
This was required because sharing from lightbox screen was not
implemented properly:
- Redundant Download: It downloaded the image in the device's storage
before picking it up to share.
- Faulty sharing 1: If a file with same name as the downloaded file was
already present in the downloaded directory, it would get picked up for
sharing instead of the file we downloaded.
- Faulty sharing 2: Zulip was unable to share to some apps.
- Hard Coded Implementation: Sharing only worked for JPEG files properly.

This implementation introduces a new function `downloadFileToCache` to
download a desired file in zulip's cache directory, this helps avoid
redundant download to local storage, and avoids faulty sharing 1, file
with same name simply overwrites previous image in cache directory.

The logic of file sharing using `ShareFileAndroid` module is improved,
now one can share any type of file, rather than just JPEG image, and the
uri shared is Content uri instead of File uri which according to me was
the reason for Faulty sharing 2.

Fixes: #4539