ios: Drop iOS 11 support; require iOS 12.1+.
A fresh look at iOS installations by platform (or the proxy we use
for that, anyway) [1] shows iOS 11 usage is 0.1% of our iOS usage,
so we're well into the time when this is OK to do.

Also, with 1.5% of users on 12 and older, iOS 12 is solidly in the
range where we'd be happy to drop it, if we find things don't work
or require effort to keep working there.

5.9% of our iOS users are on iOS <= 13, so graceful degradation at
that level is fine.

[1] https://chat.zulip.org/#narrow/stream/48-mobile/topic/platform.20versions/near/1164857