miscStyles: Use `size` prop instead of `fontSize` style for icon size.
Use the icon component's `size` prop instead of `fontSize`. Like we
did in 9a502e88b; see discussion there.

It looks like the `size={18}` wasn't doing anything: RNVI's
implementation is such that the effect of passing a `fontSize` style
is that anything passed for `size` is ignored; see
  https://github.com/oblador/react-native-vector-icons/blob/v7.0.0/lib/create-icon-set.js#L78.

Also, remove `width` and `height` from the styles. With these set to
`auto` (their default), React Native will calculate an appropriate
width and height based on the element's content; see
  https://reactnative.dev/docs/flexbox#width-and-height. Although we
*do* want those to be 24, setting them explicitly like this is a bit
belt-and-suspenders.