sentry: Clean up and document handling of sourcemap-upload key
We've had these files in version control, but with an essential
piece of them excised.  That's meant I've had a slightly silly step
in my release process, where I have a local commit that just
supplies an actual token here, and I rebase or cherry-pick that
onto each new release version before building for publication.

Instead of that process, remove and ignore those files, so I or
another release manager can just keep them sitting around in the
worktree.  (This isn't nearly as sensitive a secret as, say, the
Android release key -- all it can do is upload some bogus release
metadata to Sentry -- so there's no need for measures like
tools/checkout-keystore.)  Then move the boring config parts of
the file to documentation, and add documentation of how to create
an appropriate auth token.