docs/release: Document handling the Sentry client key
This looks kind of silly, but it's how I've been handling this for
quite some time.  (Well, together with handling the sentry.properties
files the same way, but that gets eliminated a couple of commits
before this one.)

I spent some time today investigating ways to automate this away, and
didn't come up with anything satisfying.  So for now just document
the existing process -- if we develop an improvement later, we can
apply it then.