jest: Use `jest-expo/ios` preset.
When we've used the `jest-expo` preset or the `react-native` preset,
the tests get run "in the standard React Native environment
(iOS)" [1]. The key piece of this is that they run with
`Platform.OS` mocked to 'ios'.

We're about to add another "project" to test Android codepaths with
`jest-expo/android`. First, though, swap out the old `jest-expo`
preset for this more explicit one so we don't have to think about
three presets.

[1] https://github.com/expo/expo/blob/master/packages/jest-expo/README.md#platforms