tools/test: Rename `android` suite to `native`.
We're about to add a `--platform` option, and we'd like it to
determine the behavior of some test suites, including this one. So,
give this an accurate but not platform-specific name.

(That's a cleaner way to do it than having the new option alter the
list of suites to run; such an interface would conflict with the
current interface that allows the runner to list the suites they
want to run.)