docs/style: Don't use React defaultProps for functional components
Originally discussed here:
  https://github.com/zulip/zulip-mobile/pull/4440#discussion_r565684617
(with a bit of additional rationale) when this question first came up.

It's come up a couple more times since then, as we've continued
converting a lot of our components to functions with Hooks:
  https://github.com/zulip/zulip-mobile/pull/4555#discussion_r605303854
  https://github.com/zulip/zulip-mobile/pull/4699#discussion_r622478674
so it's definitely time to have it written down in a proper place.