redux-persist: Make early `return` explicit.
It looks like the rest of this callback is only meant to run if
there hasn't been an error.

In practice, I think most of it wasn't running anyway, since there
would have been an error like "TypeError: Cannot read property
'filter' of undefined" on the `allKeys.filter(...)` just below.

Good to be explicit, though.