android: Remove meaningless attribute from camera's `uses-permission` tag.
`android:required` is not a valid attribute for `uses-permission`.
See docs:
  https://developer.android.com/guide/topics/manifest/uses-permission-element

And indeed it empirically doesn't seem to have had the desired effect
of making the app available on devices lacking a camera: see #4722.

In the next commit, we'll make a different change that the documentation
says should actually have that effect.