android: Make camera hardware requirement false.
Google Play filters which app to display to a user based on `uses-sdk`,
`uses-permission` and `uses-feature` tags.

In case `uses-feature` is not declared `uses-permission` is used to do
this filtering, and by default it treats the permission it needs as a
compulsory requirement for the app. Hence our app will not show in play
store for devices without the hardware of concern.

With `uses-feature` we override the filtering done by `uses-permission`
and by setting the required attribute to false we say that our app isn't
dependent on the hardware of concern to work properly (camera in case of
this commit).

relevant documentation:
- https://developer.android.com/guide/topics/manifest/uses-feature-element

Fixes: #4722