sharing: Move `FLAG_ACTIVITY_LAUNCHED_FROM_HISTORY` check to `onCreate`.
This conditional existed to fix a bug where a quit app, that originally
was started with SEND_INTENT, when restarted from the recent app drawer
reused the SEND_INTENT (thereby opening on sharing ui once again).

introduced at: 70e9ad9d47ae2e192558cb4caa9ae8781251b926

`onCreate` seems like a good place to have it compared to `handleSend`
as it simplifies the code and we avoid the cases where we need to check
for intents coming from other places (such as `onNewIntent`).