ServerCompatBanner: Add, to nag users connected to servers <2.0.0.
This would be around 2-4% of our users, as of 2021-03 [1].

As Greg points out [2], we'd like to try this out on a small number
of users and gather any feedback, before expanding it to more users,
since it's the kind of message that can make people unhappy. So,
even though Zulip's release lifecycle doc [3] says 2.1.0 is the
threshold, we'd like to start using that threshold only after the
code for this threshold has been out for at least a couple of weeks.

And at some time in the future, we'll likely settle on a threshold
that's not hard-coded at all; see
  https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/compatibility.20documentation/near/1174966.

[1] https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/emoji.20list/near/1136563
[2] https://github.com/zulip/zulip-mobile/pull/4750#issuecomment-848350317
[3] https://zulip.readthedocs.io/en/4.2/overview/release-lifecycle.html